django2/requirements.txt
django
whitenoise
gunicorn
django-bootstrap4
PyMySQL
django-stdimage
MySQL
